+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/manager/OldestStreamSelector.java
@@ -1,10 +1,12 @@
 package uk.gov.justice.services.event.sourcing.subscription.manager;
 
-import java.util.Optional;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
 import uk.gov.justice.services.event.buffer.core.repository.subscription.LockedStreamStatus;
 import uk.gov.justice.services.event.buffer.core.repository.subscription.NewStreamStatusRepository;
+
+import java.util.Optional;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
 
 @ApplicationScoped
 public class OldestStreamSelector implements StreamSelector {