+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/manager/StreamSelector.java
@@ -1,7 +1,8 @@
 package uk.gov.justice.services.event.sourcing.subscription.manager;
 
-import java.util.Optional;
 import uk.gov.justice.services.event.buffer.core.repository.subscription.LockedStreamStatus;
+
+import java.util.Optional;
 
 public interface StreamSelector {